Separating request auth, server and route in different files to better understanding.
